API

ObservationDims.ObsArrangement — Type

type ObsArrangement

Defines the orientation of data that is expected by a function

ObservationDims.obs_arrangement — Function

obs_arrangement(f::Type{<:Function}) -> ObsArrangement

Specify the observation arrangement trait of a function f.

ObservationDims.organise_obs — Function

organise_obs(f, data; obsdim=_default_obsdim(data))
organise_obs(::ObsArrangement, data; obsdim=_default_obsdim(data))

Organise the data according to the ObsArrangement expected by f.

Arguments

f: the function or method needing the data in a certain orientation

data: the data to transform

obsdim: the dimension of the observations, resorts to _default_obsdim when not specified.

ObsArrangements

ObservationDims.SingleObs — Type

SingleObs <: ObsArrangement

The data consists of a single observation regardless of dimension.

ObservationDims.IteratorOfObs — Type

IteratorOfObs <: ObsArrangement

The data consists of an iterator over multiple observations.

ObservationDims.ArraySlicesOfObs — Type

ArraySlicesOfObs{D} <: ObsArrangement

The data consists of a multi-dimensional array where the observations are along dimension D.

ObservationDims.MatrixRowsOfObs — Type

MatrixRowsOfObs <: ObsArrangement

A special case of ArraySlicesOfObs where the observations are along the first dimension.

ObservationDims.MatrixColsOfObs — Type

MatrixColsOfObs <: ObsArrangement

A special case of ArraySlicesOfObs where the observations are along the second dimension.